Adds in a function that changes a view to a pdf
